feat: ISBN→Long, UI overhaul, dark mode, image picker, extended filters
ISBN / persistence
- Llibre.ISBN changed Integer → Long (14-digit support)
- ServerConect: BIGINT PRIMARY KEY, setLong/getLong
- checkLlibre: countDig accepts long, validates == 14 digits

UI / theming (UITheme.java — new file)
- Centralised color + font constants via UITheme
- Nimbus L&F configured in Ejecutable with custom palette
- UITheme.setDark(boolean): full dark/light palette swap + UIManager update
- MostrarBibliotecaPanel.applyTheme(): re-applies colors + re-styles buttons after updateComponentTreeUI

Layout
- MainFramePanel: no menu bar, resizable, MAXIMIZED_BOTH, min 800×500
- MostrarBibliotecaPanel: BorderLayout table + left filter sidebar; Sortir pinned outside scroll; auto-resize columns
- GuardarLlibresDialogo: DO_NOTHING_ON_CLOSE; Guardar/Cancel·lar stacked in left margin
- DetallesLlibrePanel: Tancar button added; no decorative header strips

Filters
- New range inputs: Any (min/max), Valoració (min/max), Preu (min/max)
- ControladorDomini.aplicarFiltres extended with 4 new Double params
- MostrarBibliotecaControl.filtrar() reads + passes all new fields
- quitarFiltros() clears all range fields

Image support
- DetallesLlibrePanel: "Seleccionar imatge" button (enabled in edit mode only)
- DetallesLlibrePanelControl: fixed image scaling (was using getWidth()=0 pre-display, now uses 200×200 constant); JFileChooser starts in configured images dir
- GuardarLlibresDialogo: "Explorar..." button next to portada field
- GuardarLlibresDialogoControl: same JFileChooser wired for new-book form

+++ src/domini/Llibre.java
@@ -2,7 +2,7 @@ package domini;
 
 public class Llibre {
 
-	private Integer ISBN;
+	private Long ISBN;
 	private String Nom;
 	private String Autor;
 	private Integer Any;
@@ -12,7 +12,7 @@ public class Llibre {
 	private Boolean Llegit;
 	private String Portada;
 
-	public Llibre(Integer isbn, String nom, String autor, Integer any, String descripcio, Double valoracio, Double preu,
+	public Llibre(Long isbn, String nom, String autor, Integer any, String descripcio, Double valoracio, Double preu,
 			Boolean llegit, String portada) {
 		this.ISBN = isbn;
 		this.Nom = nom;
@@ -25,11 +25,11 @@ public class Llibre {
 		this.Portada = portada;
 	}
 
-	public Integer getISBN() {
+	public Long getISBN() {
 		return ISBN;
 	}
 
-	public void setISBN(Integer iSBN) {
+	public void setISBN(Long iSBN) {
 		ISBN = iSBN;
 	}
